Move simple_memory_allocator to a new folder
