+++ tensorflow/lite/micro/BUILD
@@ -123,33 +123,18 @@ cc_library(
 )
 
 cc_library(
-    name = "ibuffer_allocator",
-    hdrs = [
-        "ibuffer_allocator.h",
-    ],
-    copts = micro_copts(),
-    deps = [
-        "//tensorflow/lite/c:common",
-        "//tensorflow/lite/core/api",
-    ],
-)
-
-cc_library(
     name = "micro_allocator",
     srcs = [
         "micro_allocation_info.cc",
         "micro_allocator.cc",
-        "simple_memory_allocator.cc",
     ],
     hdrs = [
         "micro_allocation_info.h",
         "micro_allocator.h",
-        "simple_memory_allocator.h",
     ],
     copts = micro_copts(),
     deps = [
         ":flatbuffer_utils",
-        ":ibuffer_allocator",
         ":memory_helpers",
         ":micro_arena_constants",
         ":micro_compatibility",
@@ -160,6 +145,8 @@ cc_library(
         "//tensorflow/lite/core/api:op_resolver",
         "//tensorflow/lite/kernels:kernel_util",
         "//tensorflow/lite/kernels/internal:compatibility",
+        "//tensorflow/lite/micro/arena_allocator:ibuffer_allocator",
+        "//tensorflow/lite/micro/arena_allocator:simple_memory_allocator",
         "//tensorflow/lite/micro/memory_planner:greedy_memory_planner",
         "//tensorflow/lite/micro/memory_planner:micro_memory_planner",
         "//tensorflow/lite/schema:schema_fbs",
@@ -355,12 +342,10 @@ cc_library(
     name = "recording_allocators",
     srcs = [
         "recording_micro_allocator.cc",
-        "recording_simple_memory_allocator.cc",
     ],
     hdrs = [
         "recording_micro_allocator.h",
         "recording_micro_interpreter.h",
-        "recording_simple_memory_allocator.h",
     ],
     copts = micro_copts(),
     deps = [
@@ -369,6 +354,7 @@ cc_library(
         ":micro_framework",
         "//tensorflow/lite/core/api",
         "//tensorflow/lite/kernels/internal:compatibility",
+        "//tensorflow/lite/micro/arena_allocator:recording_simple_memory_allocator",
     ],
 )
 
@@ -451,31 +437,6 @@ cc_test(
         ":recording_allocators",
         ":test_helpers",
         "//tensorflow/lite/core/api",
-        "//tensorflow/lite/micro/testing:micro_test",
-    ],
-)
-
-cc_test(
-    name = "simple_memory_allocator_test",
-    srcs = [
-        "simple_memory_allocator_test.cc",
-    ],
-    deps = [
-        ":micro_framework",
-        ":test_helpers",
-        "//tensorflow/lite/micro/testing:micro_test",
-    ],
-)
-
-cc_test(
-    name = "recording_simple_memory_allocator_test",
-    srcs = [
-        "recording_simple_memory_allocator_test.cc",
-    ],
-    deps = [
-        ":micro_framework",
-        ":recording_allocators",
-        ":test_helpers",
         "//tensorflow/lite/micro/testing:micro_test",
     ],
 )
