+++ tensorflow/lite/micro/micro_allocation_info_test.cc
@@ -16,7 +16,7 @@ limitations under the License.
 #include "tensorflow/lite/micro/micro_allocation_info.h"
 
 #include "tensorflow/lite/micro/micro_error_reporter.h"
-#include "tensorflow/lite/micro/simple_memory_allocator.h"
+#include "tensorflow/lite/micro/arena_allocator/simple_memory_allocator.h"
 #include "tensorflow/lite/micro/test_helpers.h"
 #include "tensorflow/lite/micro/testing/micro_test.h"
