+++ tensorflow/lite/micro/recording_micro_allocator.h
@@ -18,7 +18,7 @@ limitations under the License.
 
 #include "tensorflow/lite/micro/compatibility.h"
 #include "tensorflow/lite/micro/micro_allocator.h"
-#include "tensorflow/lite/micro/recording_simple_memory_allocator.h"
+#include "tensorflow/lite/micro/arena_allocator/recording_simple_memory_allocator.h"
 
 namespace tflite {
